Extend handling of GRANT exception to REVOKE (#33) (#44)
See earlier commits working around limitations of the SQL parser
which cause us to fail at parsing some GRANT statements:

https://github.com/pingcap/parser/issues/857

We don't need to support them, so this commit extends the hacks around
the problem by detecting the error and ignoring it. This is ugly and
must be fixed... once we have additional cycles.